fuelgh/requirements/dev.txt
amqp==2.6.1
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-celery-results==2.0.1
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
flower==0.9.7
humanize==3.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
openpyxl==3.0.6
packaging==20.9
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
wget==3.2
